First pass at adding events in loose xaml
Since we can't add the events directly, I've created an attached
property to map events to commands. Those commands then get
automatically set in the CommandBindings to look for methods to run in
the data context.

It only works if the DataContext is set to an ExpandoDependencyObject
right now. The DataContext can register methods using the "_Executed"
and "_CanExecute" conventions. If they do, those commands get
automatically picked up.