Catch ValueConverter and EventHandler Errors
There are more types of errors that I would like to log. To make logging
easier, I've turned the logging function into a static helper that
requires a reference to an App. This is not a good practice, but it
allows the ValueConverters and EventHandlers (attached with the
RoutedEventCommands) to be caught automatically.