Include ChangeHandlers for ExpandoDependencyObject
Previously, I had used app.AddHandler to allow for propertychange
notifiers on the ExpandoDependencyObject. But that had two issues:
- from a developer perspective, it wasn't very useful - you had to use
one method and then check the name: "Red", "Green", etc.
- from an interactive coding perspective, it wasn't very useful - if the
ExpandoDependencyObject was created in one file and the handler was
added in another, you would end up with slowdown as changes to the file
would continue to add more and more handlers, without ever removing the
earlier ones.