Redo RoutedEventCommands
This redo removes the dependency from RoutedEventCommandCollection on
whatever element contains it, allowing multiple elements to share the
same RoutedEventCommands. It makes it possible to change entire
collections or just single elements with them and still have it
registered.

As a side effect, it also requires explicit command bindings to be added
not just for buttons and menu items, but now also for
RoutedEventCommands. This extra explicitness makes RoutedEventCommands
more general and less tied to Ruby.