Introduce DataContextHandlerExtension
And just like that, most of my work on RoutedEventCommands is obsolete.
With the ability to add EventHandlers through MarkupExtensions, I can
use a syntax much closer to the syntax already expected, where you add a
handler directly to the element. I still need the
DataContextCommandBindingExtension to work for actual commands, but I no
longer need to hookup commands just to deal with mouse button changes.

I'll leave them in for now - it's still useful since it lets me set
EventHandlers in a StaticResource / Style.